+++ tests/locale_babel.py
@@ -3,9 +3,9 @@ import babel
 
 from decimal import Decimal, ROUND_UP
 from unittest import TestCase
-from wtforms.form import _unset_value
 from wtforms import Form
-from wtforms.fields.core import DecimalField
+from wtforms.fields import DecimalField
+from wtforms.utils import unset_value
 
 
 class DummyPostData(dict):
@@ -18,11 +18,15 @@ class DummyPostData(dict):
 
 class TestLocaleDecimal(TestCase):
     class F(Form):
-        LOCALES = ['hi_IN', 'en_US']
+        class Meta:
+            locales = ['hi_IN', 'en_US']
         a = DecimalField(use_locale=True)
 
-    def _format_test(self, expected, val, LOCALES=_unset_value):
-        form = self.F(LOCALES=LOCALES, a=Decimal(val))
+    def _format_test(self, expected, val, locales=unset_value):
+        meta = None
+        if locales is not unset_value:
+            meta = {'locales': locales}
+        form = self.F(meta=meta, a=Decimal(val))
         self.assertEqual(form.a._value(), expected)
 
     def test_typeerror(self):
@@ -51,8 +55,11 @@ class TestLocaleDecimal(TestCase):
         self._format_test('-5,2', neg, ['de_DE'])
         self._format_test("-12'345.2", '-12345.2', ['de_CH'])
 
-    def _parse_test(self, raw_val, expected, LOCALES=_unset_value):
-        form = self.F(DummyPostData(a=raw_val), LOCALES=LOCALES)
+    def _parse_test(self, raw_val, expected, locales=unset_value):
+        meta = None
+        if locales is not unset_value:
+            meta = {'locales': locales}
+        form = self.F(DummyPostData(a=raw_val), meta=meta)
         if not form.validate():
             raise AssertionError(
                 'Expected value %r to parse as a decimal, instead got %r' % (
@@ -61,8 +68,11 @@ class TestLocaleDecimal(TestCase):
             )
         self.assertEqual(form.a.data, expected)
 
-    def _fail_parse(self, raw_val, expected_error, LOCALES=_unset_value):
-        form = self.F(DummyPostData(a=raw_val), LOCALES=LOCALES)
+    def _fail_parse(self, raw_val, expected_error, locales=unset_value):
+        meta = None
+        if locales is not unset_value:
+            meta = {'locales': locales}
+        form = self.F(DummyPostData(a=raw_val), meta=meta)
         assert not form.validate()
         self.assertEqual(form.a.errors[0], expected_error)