+++ wtforms/ext/appengine/fields.py
@@ -2,7 +2,6 @@ from __future__ import unicode_literals
 
 import decimal
 import operator
-import warnings
 
 from wtforms import fields, widgets
 from wtforms.compat import text_type, string_types
@@ -31,14 +30,11 @@ class ReferencePropertyField(fields.SelectFieldBase):
     widget = widgets.Select()
 
     def __init__(self, label=None, validators=None, reference_class=None,
-                 label_attr=None, get_label=None, allow_blank=False,
+                 get_label=None, allow_blank=False,
                  blank_text='', **kwargs):
         super(ReferencePropertyField, self).__init__(label, validators,
                                                      **kwargs)
-        if label_attr is not None:
-            warnings.warn('label_attr= will be removed in WTForms 1.1, use get_label= instead.', DeprecationWarning)
-            self.get_label = operator.attrgetter(label_attr)
-        elif get_label is None:
+        if get_label is None:
             self.get_label = lambda x: x
         elif isinstance(get_label, string_types):
             self.get_label = operator.attrgetter(get_label)