+++ CHANGES.rst
@@ -1,12 +1,15 @@
 WTForms Changelog
 =================
 
-Version 1.1
+Version 2.0
 -----------
 Not yet released.
 
+- Add new `class Meta` paradigm for much more powerful customization 
+  of WTForms.
 - Move i18n into core. Deprecate `wtforms.ext.i18n`.
-- Fix issue rendering SelectFields with value=True
+- Move CSRF into core. Deprecate `wtforms.ext.csrf`.
+- Fix issue rendering SelectFields with ``value=True``
 - Make `DecimalField` able to use babel locale-based number formatting.
 - Drop Python 3.2 support (Python3 support for 3.3+ only)